hopefully works

+++ server/src/main/java/net/guides/springboot2/crud/repository/DeduceriRepository.java
@@ -9,5 +9,5 @@ import net.guides.springboot2.crud.model.Deduceri;
 @Repository
 public interface DeduceriRepository extends JpaRepository<Deduceri, Integer> {
 	@Query(value = "SELECT d FROM Deduceri d WHERE d.dela <= ?1 AND ?1 <= d.panala")
-	public Deduceri getDeducereBySalariu(float salariu);
+	public Deduceri getDeducereBySalariu(int salariu);
 }
